requirements.txt
pep8
pylint
django
django-debug-toolbar
djangorestframework
django-ckeditor
psycopg2
unidecode
pillow